refactor(controller): extract deployment builder to deployment_builder.go
Part 6/7 of splitting the god-controller into per-concern files.
Tracks #311.

Move constructDeployment — the 170-line function that assembles the
inference pods Deployment from the InferenceService/Model pair and
the per-runtime backend — into its own deployment_builder.go sibling.
Also move the pod- and container-level security-context helpers that
are exclusive to the inference pod; initContainerSecurityContext
stays behind because it is shared with the storage builders.

Moved symbols:
  inferPodSecurityContext
  inferContainerSecurityContext
  (*InferenceServiceReconciler).constructDeployment

Pure relocation — no signature changes, no behavior changes. Callers
in the same package continue to resolve the unqualified names.

Line deltas (based on main; if earlier parts merge first the
controller file drops further on rebase):
  inferenceservice_controller.go: 1,567 → 1,375 (−192)
  deployment_builder.go:                  0 → 229 (new)

Verified: go vet clean, golangci-lint 0 issues, make test green,
internal/controller coverage unchanged at 83.1%.
